Refactored - Created different packages and moved respective classes
