+++ build.gradle
@@ -19,6 +19,14 @@ dependencies {
     implementation('com.google.inject:guice:5.0.1')
 }
 
-test {
-    useTestNG()
+
+task runApiTests(type: Test) {
+    outputs.upToDateWhen { false }  //To avoid cache
+    useTestNG{
+        parallel = "methods"
+        threadCount Integer.parseInt(System.getProperty("sessions", "2"))
+        includeGroups System.getProperty("tags", "api")
+        useDefaultListeners true
+        outputDirectory = file("$buildDir/" + System.getProperty('tags', 'NONE'))
+    }
 }
