+++ src/test/java/com/dummy/api/tests/BaseTest.java
@@ -1,0 +1,7 @@
+package com.dummy.api.tests;
+
+import org.testng.annotations.Guice;
+
+@Guice
+public class BaseTest {
+}
