+++ src/test/java/com/dummy/api/tests/users/GetListOfUsersTests.java
@@ -1,6 +1,6 @@
-package com.dummy.api.tests;
-import com.dummy.api.user.UserClient;
-import com.dummy.api.user.get.GetUsersListResponse;
+package com.dummy.api.tests.users;
+import com.dummy.api.endpoints.users.UserClient;
+import com.dummy.api.endpoints.users.get.GetUsersListResponse;
 import org.testng.Assert;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
